requirements.txt
asgiref==3.8.1
awscli==1.38.22
awsebcli==3.21.0
blessed==1.20.0
boto3==1.37.28
botocore==1.37.28
cement==2.10.14
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
Django==4.2.20
django-storages==1.14.6
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
docutils==0.16
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
packaging==24.2
pathspec==0.10.1
pyasn1==0.6.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
rsa==4.7.2
s3transfer==0.11.4
semantic-version==2.10.0
six==1.17.0
sqlparse==0.5.3
termcolor==2.5.0
typing_extensions==4.12.2
urllib3==1.26.20
wcwidth==0.2.13
